+++ src/consts/index.js
@@ -1,0 +1,2 @@
+const undefinedBrand = null; // null - is undefined brand
+export const BRANDS_WITH_MULTIPLE_MASKS = [undefinedBrand, "mir"];