+++ src/services/card-info/index.js
@@ -2,6 +2,7 @@ import Card from "./card";
 import banks from "./banks";
 import brands from "./brands";
 import prefixes from "./prefixes";
+import { generateBlock, generateMask } from "../../utils/helpers";
 
 export default class CardInfo extends Card {
     constructor(numberSource = "", options) {
@@ -61,6 +62,18 @@ export default class CardInfo extends Card {
             this.numberGaps,
             this.numberLengths
         );
+
+        this.allBlocks = this._getAllBlocks(
+            this.numberGaps,
+            this.numberLengths
+        );
+
+        this.allMasks = this._getAllMasks(
+            this.options.maskDigitSymbol,
+            this.options.maskDelimiterSymbol,
+            this.allBlocks
+        );
+
         this.numberMask = this._getMask(
             this.options.maskDigitSymbol,
             this.options.maskDelimiterSymbol,
@@ -84,7 +97,7 @@ export default class CardInfo extends Card {
 
     _getBrand(number) {
         const brs = [];
-        Object.values(brands).forEach(brand => {
+        Object.values(brands).forEach((brand) => {
             if (brand.pattern.test(number)) brs.push(brand);
         });
         if (brs.length === 1) return brs[0];
@@ -147,6 +160,15 @@ export default class CardInfo extends Card {
         return blocks.reverse();
     }
 
+    _getAllBlocks(numberGaps, numberLengths) {
+        let result = [];
+        for (let numberLength of numberLengths) {
+            const block = generateBlock(numberGaps, numberLength);
+            result.push(block);
+        }
+        return result;
+    }
+
     _getMask(maskDigitSymbol, maskDelimiterSymbol, numberBlocks) {
         let mask = "";
         for (let i = 0; i < numberBlocks.length; i++) {
@@ -155,6 +177,16 @@ export default class CardInfo extends Card {
                 Array(numberBlocks[i] + 1).join(maskDigitSymbol);
         }
         return mask;
+    }
+
+    _getAllMasks(maskDigitSymbol, maskDelimiterSymbol, allBlocks) {
+        const masks = [];
+        for (const numberBlocks of allBlocks) {
+            const mask = generateMask(numberBlocks);
+
+            masks.push(mask);
+        }
+        return masks;
     }
 
     _getNumberNice(number, numberGaps) {